Move database tests to src
